# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/carreras": {
            "get": {
                "summary": "Get list of carreras",
                "operationId": "24f82a4261c194f035a7ad6c076d6e61",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new carrera",
                "operationId": "3430b5198844e6744484d7e1b441db8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Fecha",
                                    "CircuitoId",
                                    "CampeonatoId"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Carrera A"
                                    },
                                    "Fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "CircuitoId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "CampeonatoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/carreras/{id}": {
            "get": {
                "summary": "Get carrera by ID",
                "operationId": "8a30939fbecca1a64e1dbcdb8a97d059",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update carrera by ID",
                "operationId": "d40721a6b68ff4af3154939207cb3a1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Fecha",
                                    "CircuitoId",
                                    "CampeonatoId"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Carrera A"
                                    },
                                    "Fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "CircuitoId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "CampeonatoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete carrera by ID",
                "operationId": "cf5360ba43551c82c0a1d58246d97560",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/circuitos": {
            "get": {
                "summary": "Get list of circuitos",
                "operationId": "56b7010805fec325ef3c243732596c82",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new circuito",
                "operationId": "d06ba5b22fd79bdbbe9490b87a3ba2d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Ubicacion",
                                    "Longitud"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Circuito A"
                                    },
                                    "Ubicacion": {
                                        "type": "string",
                                        "example": "Ubicacion A"
                                    },
                                    "Longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/circuitos/{id}": {
            "get": {
                "summary": "Get circuito by ID",
                "operationId": "58710e4f05637013355740aa274ad03d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update circuito by ID",
                "operationId": "61c63912c8bec998dce727b996b9e0c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Ubicacion",
                                    "Longitud"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Circuito A"
                                    },
                                    "Ubicacion": {
                                        "type": "string",
                                        "example": "Ubicacion A"
                                    },
                                    "Longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete circuito by ID",
                "operationId": "c5c9f6a2df9384ca2194b0bea3948dd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}